Add files via upload
In this project, I queried the Chinook Database. The Chinook Database holds information about a music store. For this project, I be assisted the Chinook team with understanding the media in their store, their customers and employees, and their invoice information.
My presentation includes; Four slides, One visualization per slide, A 1-2 sentence explanation of each slide.
The SQL query is used to create the data used in the visualization.